Finished Assignment 7 Milestone 4
Removed .class files from the git tracking.
Added ability of FacePamphletDatabase to remove a profile from all other profiles' list of friends.
Changed FacePamphletProfile to use a string to set the image, not require a GImage. Subsequently, add error handling for invalid image filenames.
Added currentProfile to FacePamphlet and FacePamphletConsole, for tracking of the current profile. Added currentProfile assignments when adding, deleting, and looking up a profile.
Added handling of status, profile, and friends to FacePamphlet and FacePamphletConsole through the profile database.